+++ network/internal/readConnection.go
@@ -1,4 +1,4 @@
-package protocol
+package internal
 
 import (
 	"context"
@@ -12,6 +12,7 @@ import (
 	"github.com/onflow/flow-go/module"
 	"github.com/onflow/flow-go/module/metrics"
 	"github.com/onflow/flow-go/network/message"
+	"github.com/onflow/flow-go/network/protocol"
 )
 
 // readConnection reads the incoming stream and calls the callback until the remote closes the stream or the context is
@@ -26,7 +27,7 @@ type readConnection struct {
 }
 
 // newReadConnection creates a new readConnection
-func newReadConnection(ctx context.Context,
+func NewReadConnection(ctx context.Context,
 	stream libp2pnetwork.Stream,
 	callback func(msg *message.Message),
 	log zerolog.Logger,
@@ -34,7 +35,7 @@ func newReadConnection(ctx context.Context,
 	maxMsgSize int) *readConnection {
 
 	if maxMsgSize <= 0 {
-		maxMsgSize = DefaultMaxUnicastMsgSize
+		maxMsgSize = protocol.DefaultMaxUnicastMsgSize
 	}
 
 	c := readConnection{
@@ -48,9 +49,9 @@ func newReadConnection(ctx context.Context,
 	return &c
 }
 
-// receiveLoop must be run in a goroutine and it continuously reads messages from the peer until
+// ReceiveLoop must be run in a goroutine and it continuously reads messages from the peer until
 // either the remote closes the stream or the context is cancelled
-func (rc *readConnection) receiveLoop(wg *sync.WaitGroup) {
+func (rc *readConnection) ReceiveLoop(wg *sync.WaitGroup) {
 
 	defer wg.Done()
 	defer rc.log.Debug().Msg("exiting receive routine")
