+++ network/internal/readSubscription.go
@@ -1,4 +1,4 @@
-package network
+package internal
 
 import (
 	"context"
@@ -23,7 +23,7 @@ type readSubscription struct {
 }
 
 // newReadSubscription reads the messages coming in on the subscription
-func newReadSubscription(ctx context.Context,
+func NewReadSubscription(ctx context.Context,
 	sub *pubsub.Subscription,
 	callback func(msg *message.Message),
 	log zerolog.Logger,
@@ -44,9 +44,9 @@ func newReadSubscription(ctx context.Context,
 	return &r
 }
 
-// receiveLoop must be run in a goroutine. It continuously receives
+// ReceiveLoop must be run in a goroutine. It continuously receives
 // messages for the topic and calls the callback synchronously
-func (r *readSubscription) receiveLoop(wg *sync.WaitGroup) {
+func (r *readSubscription) ReceiveLoop(wg *sync.WaitGroup) {
 
 	defer wg.Done()
 	defer r.log.Debug().Msg("exiting receive routine")
